+++ src/api/interfaces.ts
@@ -7,11 +7,12 @@ export type ReadyEvent = {
 
 export type ErrorEvent = {
 	type: string;
-	error: Error;
+	message: string;
 };
 
 export type FatalEvent = {
-	error: Error;
+	message: string;
+	log?: string;
 };
 
 export type InvalidEvent = {