+++ src/core/create_routes.ts
@@ -5,17 +5,22 @@ import { Route } from '../interfaces';
 
 export default function create_routes({ files } = { files: glob.sync('**/*.*', { cwd: locations.routes(), dot: true, nodir: true }) }) {
     const routes: Route[] = files
-        .filter((file: string) => !/(^|\/|\\)_/.test(file))
+        .filter((file: string) => !/(^|\/|\\)(_(?!error\.html)|\.(?!well-known))/.test(file))
         .map((file: string) => {
-            if (/(^|\/|\\)(_|\.(?!well-known))/.test(file)) return;
-
             if (/]\[/.test(file)) {
                 throw new Error(`Invalid route ${file} — parameters must be separated`);
             }
 
+            if (file === '4xx.html' || file === '5xx.html') {
+                throw new Error('As of Sapper 0.14, 4xx.html and 5xx.html should be replaced with _error.html');
+            }
+
             const base = file.replace(/\.[^/.]+$/, '');
             const parts = base.split('/'); // glob output is always posix-style
-            if (parts[parts.length - 1] === 'index') parts.pop();
+            if (/^index(\..+)?/.test(parts[parts.length - 1])) {
+                const part = parts.pop();
+                if (parts.length > 0) parts[parts.length - 1] += part.slice(5);
+            }
 
             return {
                 files: [file],
@@ -30,8 +35,8 @@ export default function create_routes({ files } = { files: glob.sync('**/*.*', {
             return !found;
         })
         .sort((a, b) => {
-            if (a.parts[0] === '4xx' || a.parts[0] === '5xx') return -1;
-            if (b.parts[0] === '4xx' || b.parts[0] === '5xx') return 1;
+            if (a.parts[0] === '_error') return -1;
+            if (b.parts[0] === '_error') return 1;
 
             const max = Math.max(a.parts.length, b.parts.length);
 
@@ -63,6 +68,7 @@ export default function create_routes({ files } = { files: glob.sync('**/*.*', {
                             (a_sub_part.content < b_sub_part.content ? -1 : 1)
                         );
                     }
+
                     // If both parts dynamic, check for regexp patterns
                     if (a_sub_part.dynamic && b_sub_part.dynamic) {
                         const regexp_pattern = /\((.*?)\)/;