+++ src/middleware.ts
@@ -1,15 +1,12 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { resolve, URL } from 'url';
+import { URL } from 'url';
 import { ClientRequest, ServerResponse } from 'http';
 import cookie from 'cookie';
-import mkdirp from 'mkdirp';
-import rimraf from 'rimraf';
 import devalue from 'devalue';
 import fetch from 'node-fetch';
 import { lookup } from './middleware/mime';
 import { locations, dev } from './config';
-import { Route, Template } from './interfaces';
 import sourceMapSupport from 'source-map-support';
 
 sourceMapSupport.install();
@@ -59,14 +56,29 @@ export default function middleware({ App, routes, store }: {
 
 	const output = locations.dest();
 
-	const client_assets = JSON.parse(fs.readFileSync(path.join(output, 'client_assets.json'), 'utf-8'));
+	let emitted_basepath = false;
 
 	const middleware = compose_handlers([
 		(req: Req, res: ServerResponse, next: () => void) => {
 			if (req.baseUrl === undefined) {
-				req.baseUrl = req.originalUrl
-					? req.originalUrl.slice(0, -req.url.length)
+				let { originalUrl } = req;
+				if (req.url === '/' && originalUrl[originalUrl.length - 1] !== '/') {
+					originalUrl += '/';
+				}
+
+				req.baseUrl = originalUrl
+					? originalUrl.slice(0, -req.url.length)
 					: '';
+			}
+
+			if (!emitted_basepath && process.send) {
+				process.send({
+					__sapper__: true,
+					event: 'basepath',
+					basepath: req.baseUrl
+				});
+
+				emitted_basepath = true;
 			}
 
 			if (req.path === undefined) {
@@ -96,7 +108,7 @@ export default function middleware({ App, routes, store }: {
 			cache_control: 'max-age=31536000'
 		}),
 
-		get_route_handler(client_assets, App, routes, store)
+		get_route_handler(App, routes, store)
 	].filter(Boolean));
 
 	return middleware;
@@ -139,24 +151,41 @@ function serve({ prefix, pathname, cache_control }: {
 	};
 }
 
-const resolved = Promise.resolve();
+function get_route_handler(App: Component, routes: RouteObject[], store_getter: (req: Req) => Store) {
+	const output = locations.dest();
 
-function get_route_handler(chunks: Record<string, string>, App: Component, routes: RouteObject[], store_getter: (req: Req) => Store) {
+	const get_chunks = dev()
+		? () => JSON.parse(fs.readFileSync(path.join(output, 'client_assets.json'), 'utf-8'))
+		: (assets => () => assets)(JSON.parse(fs.readFileSync(path.join(output, 'client_assets.json'), 'utf-8')));
+
 	const template = dev()
 		? () => fs.readFileSync(`${locations.app()}/template.html`, 'utf-8')
 		: (str => () => str)(fs.readFileSync(`${locations.dest()}/template.html`, 'utf-8'));
 
-	function handle_route(route: RouteObject, req: Req, res: ServerResponse) {
-		req.params = route.params(route.pattern.exec(req.path));
+	const error_route = routes.find((route: RouteObject) => route.error);
+
+	function handle_route(route: RouteObject, req: Req, res: ServerResponse, status = 200, error: Error | string = null) {
+		req.params = error
+			? {}
+			: route.params(route.pattern.exec(req.path));
 
 		const handlers = route.handlers[Symbol.iterator]();
 
 		function next() {
+			const chunks: Record<string, string> = get_chunks();
+
 			try {
 				const { value: handler, done } = handlers.next();
 
 				if (done) {
-					handle_error(req, res, 404, 'Not found');
+					if (route.error) {
+						// there was an error rendering the error page!
+						res.statusCode = status;
+						res.end(error instanceof Error ? error.message : error);
+					} else {
+						handle_route(error_route, req, res, 404, 'Not found');
+					}
+
 					return;
 				}
 
@@ -168,7 +197,7 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 					// preload main.js and current route
 					// TODO detect other stuff we can preload? images, CSS, fonts?
 					const link = []
-						.concat(chunks.main, chunks[route.id])
+						.concat(chunks.main, chunks[route.id] || chunks._error) // TODO this is gross
 						.filter(file => !file.match(/\.map$/))
 						.map(file => `<${req.baseUrl}/client/${file}>;rel="preload";as="script"`)
 						.join(', ');
@@ -178,8 +207,13 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 					const store = store_getter ? store_getter(req) : null;
 					const props = { params: req.params, query: req.query, path: req.path };
 
+					if (route.error) {
+						props.error = error instanceof Error ? error : { message: error };
+						props.status = status;
+					}
+
 					let redirect: { statusCode: number, location: string };
-					let error: { statusCode: number, message: Error | string };
+					let preload_error: { statusCode: number, message: Error | string };
 
 					Promise.resolve(
 						mod.preload ? mod.preload.call({
@@ -187,7 +221,7 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 								redirect = { statusCode, location };
 							},
 							error: (statusCode: number, message: Error | string) => {
-								error = { statusCode, message };
+								preload_error = { statusCode, message };
 							},
 							fetch: (url: string, opts?: any) => {
 								const parsed = new URL(url, `http://127.0.0.1:${process.env.PORT}${req.baseUrl ? req.baseUrl + '/'  :''}`);
@@ -227,7 +261,7 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 						store
 						}, req) : {}
 					).catch(err => {
-						error = { statusCode: 500, message: err };
+						preload_error = { statusCode: 500, message: err };
 					}).then(preloaded => {
 						if (redirect) {
 							res.statusCode = redirect.statusCode;
@@ -237,8 +271,8 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 							return;
 						}
 
-						if (error) {
-							handle_error(req, res, error.statusCode, error.message);
+						if (preload_error) {
+							handle_route(error_route, req, res, preload_error.statusCode, preload_error.message);
 							return;
 						}
 
@@ -276,11 +310,13 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 							.replace('%sapper.head%', `<noscript id='sapper-head-start'></noscript>${head}<noscript id='sapper-head-end'></noscript>`)
 							.replace('%sapper.styles%', (css && css.code ? `<style>${css.code}</style>` : ''));
 
+						res.statusCode = status;
 						res.end(page);
 
 						if (process.send) {
 							process.send({
 								__sapper__: true,
+								event: 'file',
 								url: req.url,
 								method: req.method,
 								status: 200,
@@ -320,6 +356,7 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 
 								process.send({
 									__sapper__: true,
+									event: 'file',
 									url: req.url,
 									method: req.method,
 									status: res.statusCode,
@@ -350,76 +387,17 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 					}
 				}
 			} catch (error) {
-				handle_error(req, res, 500, error);
+				if (route.error) {
+					// there was an error rendering the error page!
+					res.statusCode = status;
+					res.end(error instanceof Error ? error.message : error);
+				} else {
+					handle_route(error_route, req, res, 500, error || 'Internal server error');
+				}
 			}
 		}
 
 		next();
-	}
-
-	const not_found_route = routes.find((route: RouteObject) => route.error === '4xx');
-	const error_route = routes.find((route: RouteObject) => route.error === '5xx');
-
-	function handle_error(req: Req, res: ServerResponse, statusCode: number, message: Error | string) {
-		res.statusCode = statusCode;
-		res.setHeader('Content-Type', 'text/html');
-
-		const error = message instanceof Error ? message : new Error(message);
-
-		const not_found = statusCode >= 400 && statusCode < 500;
-
-		const route = not_found
-			? not_found_route
-			: error_route;
-
-		function render_page({ head, css, html }) {
-			const page = template()
-				.replace('%sapper.base%', `<base href="${req.baseUrl}/">`)
-				.replace('%sapper.scripts%', `<script>__SAPPER__={baseUrl: "${req.baseUrl}"}</script><script src='${req.baseUrl}/client/${chunks.main}'></script>`)
-				.replace('%sapper.html%', html)
-				.replace('%sapper.head%', `<noscript id='sapper-head-start'></noscript>${head}<noscript id='sapper-head-end'></noscript>`)
-				.replace('%sapper.styles%', (css && css.code ? `<style>${css.code}</style>` : ''));
-
-			res.end(page);
-		}
-
-		function handle_notfound() {
-			const title: string = not_found
-				? 'Not found'
-				: `Internal server error: ${error.message}`;
-
-			render_page({ head: '', css: null, html: title });
-		}
-
-		if (route) {
-			const handlers = route.handlers[Symbol.iterator]();
-
-			function next() {
-				const { value: handler, done } = handlers.next();
-
-				if (done) {
-					handle_notfound();
-				} else if (handler.type === 'page') {
-					render_page(handler.module.render({
-						status: statusCode,
-						error
-					}, {
-						store: store_getter && store_getter(req)
-					}));
-				} else {
-					const handle_method = mod[method_export];
-					if (handle_method) {
-						handle_method(req, res, next);
-					} else {
-						next();
-					}
-				}
-			}
-
-			next();
-		} else {
-			handle_notfound();
-		}
 	}
 
 	return function find_route(req: Req, res: ServerResponse) {
@@ -427,7 +405,7 @@ function get_route_handler(chunks: Record<string, string>, App: Component, route
 			if (!route.error && route.pattern.test(req.path)) return handle_route(route, req, res);
 		}
 
-		handle_error(req, res, 404, 'Not found');
+		handle_route(error_route, req, res, 404, 'Not found');
 	};
 }
 
@@ -449,10 +427,6 @@ function compose_handlers(handlers: Handler[]) {
 
 		go();
 	};
-}
-
-function read_json(file: string) {
-	return JSON.parse(fs.readFileSync(file, 'utf-8'));
 }
 
 function try_serialize(data: any) {